Move configure-feature
