+++ src/configure-feature.js
